Usability improvements for hello-vulkan sample
Move member functions to a cpp file
Remove redundant functions and calls
Report validation layer messages to logcat
Auto-disable validation layers when not found
Feature support checks to avoid validation failures
Extra logging to help with debugging when using the app
