+++ hello-vulkan/app/src/main/cpp/CMakeLists.txt
@@ -31,7 +31,8 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++17 -Wall")
 add_definitions(-DVK_USE_PLATFORM_ANDROID_KHR=1)
 
 add_library(${PROJECT_NAME} SHARED
-    vk_main.cpp)
+    vk_main.cpp
+    hellovk.cpp)
 
 # add lib dependencies
 target_link_libraries(${PROJECT_NAME} PUBLIC
